add support for uploading report to s3
