+++ configs/hidra/exporter.yml
@@ -2,3 +2,5 @@ log_level: debug
 http_server:
   listen_address: :19090
 samples_path: ./configs/hidra/samples
+report:
+  enabled: true
